# openapi.yaml
openapi: 3.0.3
info:
  title: Armada Convoy API
  description: REST API for the Armada convoy and social app
  version: 1.0.0
  contact:
    name: Armada Team
    email: support@armada.app
servers:
  - url: https://convoy-api.onrender.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      description: Create a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9_]+$'
                  example: johndoe
                password:
                  type: string
                  minLength: 8
                  example: SecurePassword123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      summary: Login user
      description: Authenticate user and return tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - emailOrUsername
                - password
              properties:
                emailOrUsername:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: SecurePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe
        avatarUrl:
          type: string
          format: uri
          example: https://example.com/avatar.jpg
        bio:
          type: string
          example: Adventure enthusiast and convoy leader
        settings:
          $ref: '#/components/schemas/UserSettings'
        stats:
          $ref: '#/components/schemas/UserStats'
        friends:
          type: array
          items:
            $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserSettings:
      type: object
      properties:
        showConvoys:
          type: string
          enum: [everyone, friends, private]
          example: friends
        showStats:
          type: string
          enum: [everyone, friends, private]
          example: friends
        showProfile:
          type: string
          enum: [everyone, friends, private]
          example: everyone
        notifications:
          type: object
          properties:
            convoyInvites:
              type: boolean
              example: true
            friendRequests:
              type: boolean
              example: true
            messages:
              type: boolean
              example: true
            achievements:
              type: boolean
              example: true
    UserStats:
      type: object
      properties:
        convoysCreated:
          type: integer
          example: 12
        convoysJoined:
          type: integer
          example: 45
        totalMiles:
          type: number
          example: 1250.5
        friendsCount:
          type: integer
          example: 23
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...